[Refactoring ⚙️] delete Comment 관련 훅 생성 및 연결
