--- src/components/PostDetail/apis/deleteComment.ts
@@ -1,13 +1,0 @@
-import { AUTH_API } from "@/apis/Api"
-
-const deleteComment = (commentId: string) => {
-  return AUTH_API.delete("/comments/delete", {
-    data: { id: commentId },
-  })
-    .then((res) => res.data)
-    .catch((error) => {
-      throw Error(error)
-    })
-}
-
-export default deleteComment
